+++ FBSDKCoreKit/FBSDKCoreKit/include/FBSDKWebDialogView.h
@@ -11,8 +11,6 @@
 #import <UIKit/UIKit.h>
 
 @protocol FBSDKWebDialogViewDelegate;
-@protocol FBSDKWebViewProviding;
-@protocol FBSDKInternalURLOpener;
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -20,9 +18,6 @@ NS_SWIFT_NAME(FBWebDialogView)
 @interface FBSDKWebDialogView : UIView
 
 @property (nonatomic, weak) id<FBSDKWebDialogViewDelegate> delegate;
-
-+ (void)configureWithWebViewProvider:(id<FBSDKWebViewProviding>)provider
-                           urlOpener:(id<FBSDKInternalURLOpener>)urlOpener;
 
 - (void)loadURL:(NSURL *)URL;
 - (void)stopLoading;