+++ FBSDKCoreKit/FBSDKCoreKitTests/FBSDKCoreKitTests-Bridging-Header.h
@@ -192,6 +192,7 @@
 #import "FBSDKViewImpressionLogger+Testing.h"
 #import "FBSDKWebDialog+Testing.h"
 #import "FBSDKWebDialogView+Testing.h"
+#import "FBSDKWebViewFactory.h"
 #import "FBSDKWindowFinding.h"
 #import "FeatureManager+Testing.h"
 #import "ImageDownloader+Testing.h"