+++ FBSDKCoreKit/FBSDKCoreKitTests/Interfaces/FBSDKWebDialogView+Testing.h
@@ -6,24 +6,17 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-#import <FBSDKCoreKit/FBSDKCoreKit.h>
-
-#import "FBSDKWebDialogView.h"
-#import "FBSDKWebViewFactory.h"
-#import "FBSDKWebViewProviding.h"
-#import "WKWebView+WebViewProtocol.h"
+#import "FBSDKWebDialogView+Internal.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
 @interface FBSDKWebDialogView (Testing) <WKNavigationDelegate>
 
-@property (class, nullable, nonatomic, readonly, strong) id<FBSDKWebViewProviding> webViewProvider;
-@property (class, nullable, nonatomic, readonly, strong) id<FBSDKURLOpener> urlOpener;
 @property (nonatomic, strong) UIButton *closeButton;
 @property (nonatomic, strong) UIActivityIndicatorView *loadingView;
 @property (nullable, nonatomic, strong) id<FBSDKWebView> webView;
 
-+ (void)reset;
++ (void)resetClassDependencies;
 
 @end